+++ karel/worlds/2x1.w
@@ -1,0 +1,4 @@
+Dimension: (2, 1)
+Karel: (1, 1) east
+BeeperBag: INFINITE
+Speed: 0.75
