+++ karel/worlds/MidpointFindingKarel.w
@@ -1,0 +1,4 @@
+Dimension: (9, 9)
+BeeperBag: INFINITE
+Karel: (1, 1) East
+Speed: 0.50
