+++ karel/worlds/SampleQuad2.w
@@ -1,0 +1,35 @@
+Dimension: (13, 9)
+Beeper: (1, 1) 1
+Beeper: (1, 2) 1
+Wall: (1, 4) South
+Wall: (2, 4) West
+Wall: (2, 5) South
+Wall: (3, 4) West
+Wall: (3, 4) South
+Wall: (4, 4) West
+Wall: (4, 5) South
+Beeper: (5, 1) 1
+Wall: (5, 5) West
+Beeper: (5, 5) 1
+Wall: (5, 6) South
+Wall: (6, 5) West
+Wall: (6, 5) South
+Wall: (7, 4) West
+Wall: (7, 4) South
+Wall: (8, 3) West
+Wall: (8, 3) South
+Wall: (9, 2) West
+Wall: (9, 2) South
+Wall: (10, 2) West
+Wall: (10, 3) South
+Wall: (11, 3) West
+Wall: (11, 4) South
+Wall: (12, 4) West
+Wall: (12, 5) South
+Beeper: (13, 1) 1
+Beeper: (13, 3) 1
+Wall: (13, 4) West
+Wall: (13, 4) South
+BeeperBag: INFINITE
+Karel: (1, 1) East
+Speed: 0.5
